// src/client/types.gen.ts
export type CreateTodoCommand = {
    /**
     * The title of the todo
     */
    title: string;
    description: string | null;
    deadline: string | null;
};